get jobs bug fix
