+++ client/src/components/History/HistoryViews/JobHistory.jsx
@@ -27,17 +27,6 @@ const JobHistory = ({ className }) => {
     };
   };
 
-  const dispatch = useDispatch();
-
-  const query = queryStringParser.parse(useLocation().search);
-
-  useEffect(() => {
-    dispatch({
-      type: 'GET_JOBS',
-      params: { offset: 0, queryString: query.query_string },
-    });
-  }, [dispatch, query.query_string]);
-
   return (
     /* CLEVER: Using `job-history` class to scope the `unread` className */
     /* FP-936: Use 'unread' class from CSS Modules stylesheet, not global one */
